% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencingHeatmap.R
\name{sequencingHeatmap}
\alias{sequencingHeatmap}
\title{Sequencing Data Heatmap Creation}
\usage{
sequencingHeatmap(sequencing, sheet, columns, inputs, method, cutoff.p = 0.4,
  baseMeanCount = 15, top)
}
\arguments{
\item{sequencing}{The path to the sequencing output spreadsheet (.xlsx), i.e. sequencing="path/to/output.xlsx"}

\item{sheet}{The sheet number in the spreadsheet with data, i.e sheet=1}

\item{columns}{The column range of data to use, i.e. columns=5:31}

\item{inputs}{The path to the directory which contains files of genes of interest, i.e.
inputs="path/to/inputs/"}

\item{method}{The method of identifying each gene. It must be either method="symbol" or method="geneid"}

\item{top}{(optional) The number of top genes to select, filtered by adjusted p-value. Can use
multiple values for one run, i.e. top=100 or top=c(100,125,150)}
}
\description{
This function allows you to create heatmaps from a sequencing output spreadsheet.
}
\details{
\strong{Formatting of an input file:} \cr
Method of clustering the results, either "gene", "sample", or "both" \cr
Title of the heatmap plot \cr
GENE1 \cr
GENE2 \cr
GENE3 \cr
... \cr
\cr
\strong{Example input file:} \cr
gene \cr
Sample Graph Title \cr
SFRP2 \cr
CD4 \cr
BRCA1 \cr
... \cr
\cr
\strong{Outputs} (found in \code{./sequencingHeatmap-output})\strong{:} \cr
PDF, TIFF, FIG, and EPS of heatmaps \cr
CSV of fold-changes \cr
* All file names correspond to those in the provided inputs directory \cr
\cr
\strong{Directory structure for inputs and resultings outputs with given args:}
\preformatted{
.
|---sequencingoutput.xlsx                 # sequencing='sequencingoutput.xlsx'
|---experiment1                      # inputs='experiment1'
|   |---genelist1.txt
|----sequencingHeatmap-output
|   |---experiment1
|       |---genelist1-top100.pdf     # top=c(100,125)
|       |---genelist1-top125.pdf     
|       |---genelist1-top100.tiff
|       |---genelist1-top125.tiff
|       |---genelist1-top100.eps
|       |---genelist1-top125.eps
|       |---genelist1-top100.fig
|       |---genelist1-top125.fig
|       |---genelist1-top100-foldChange.csv
|       |---genelist1-top125-foldChange.csv
|       |---sequencingoutput-genelist1-top100-truncated.csv
|       |---sequencingoutput-genelist1-top125-truncated.csv}
}
\examples{
\dontrun{

library(sequencingHeatmap)
sequencingHeatmap(sequencing="sequencingoutput.xlsx", sheet=1, columns=5:14,
inputs="experiment1", method="symbol", top=c(100,125,150))
}
}

